requirements.txt
# Requirements for OCR-CarPlate project
ultralytics>=8.0.0
python-dotenv>=0.19.0
torch>=1.13.0
opencv-python>=4.5.0
pytesseract>=0.3.9
numpy>=1.19.0